Queue using two stacks
The code implements a queue using two stacks. It provides functionalities to enqueue elements into the queue, dequeue elements from the queue, and print the front element without removing it. It reads user queries where:

1 x denotes enqueuing element x.
2 denotes dequeuing an element.
3 denotes printing the front element of the queue without removing it.
Internally, it uses two stacks: one for enqueueing elements and another for dequeueing. When dequeue is called and the dequeue stack is empty, it transfers elements from the enqueue stack to the dequeue stack to maintain the order of elements. This process allows enqueue and dequeue operations to work efficiently.